dont ssr web3 things
